Serve most of the traffic from service worker except JSON files sometimes
Now everything, including entry points, are served from the service
worker (when applicable obviously). As a result the entire website
is offline-browsable for the pages previously visited, including when
you reload an HTML page you didn’t previously visited if you previously
hydrated it (it’s a subtle but expected improvement).