Detect moved/renamed articles
Probably not an usual operation for in-force laws (except possibly for
re-codifications), but an usual operation for law proposals.

Note that only identical contents are recognised as moved/renamed, and
a nice feature would be to have a similarity function to classify
moved/renamed articles.

Note also that the model for moved articles is a set of articles of the
first text to a set of articles in the second text: sometimes a common
content can correspond to multiple moved articles, and there should be
a post-treatment function to link moved articles. Obviously the easy
case is when each set has only one element.